+++ dlls/d3d9/d3d9_private.h
@@ -482,7 +482,7 @@ typedef struct IDirect3DVertexDeclaration9Impl {
   LONG    ref;
 
   D3DVERTEXELEMENT9 *elements;
-  size_t element_count;
+  UINT element_count;
 
   /* IDirect3DVertexDeclaration9 fields */
   IWineD3DVertexDeclaration *wineD3DVertexDeclaration;
